Workaround for SCTP port 36412
When port 36412/SCTP is explicitly bound in docker-compose.yaml,
an eNB can reach it from an external IP, from a Docker bridge
(172.18.0.1) and from an internal IP (172.18.0.3).  However, a
container within the bridged network (srsenb in our case) cannot
reach it by internal IP.  When the port is removed, it is reachable.
This might be a bug in docker-compose.